+++ R/path.R
@@ -34,6 +34,7 @@ standardise_one_path <- function (path, check = TRUE) {
     ext <- tolower(tools::file_ext(path))
     return(
       switch(ext,
+        zip = ,
         bz2 = ,
         xz = {
           close(con)
@@ -57,7 +58,11 @@ standardise_one_path <- function (path, check = TRUE) {
     bz2 = bzfile(path, ""),
     xz = xzfile(path, ""),
     zip = zipfile(path, ""),
-    path
+    if (!has_trailing_newline(path)) {
+      file(path)
+    } else {
+      path
+    }
   )
 }