+++ inst/bench/all_character-benchmark.R
@@ -1,5 +1,7 @@
-rows <- 1e6
-cols <- 25
+args <- commandArgs(trailingOnly = TRUE)
+str(args)
+rows <- as.integer(args[[1]])
+cols <- as.integer(args[[2]])
 
 set.seed(42)
 RNGversion("3.5.3")
@@ -33,36 +35,6 @@ vroom_write(data, file, "\t")
 
 desc <- c("setup", "read", "print", "head", "tail", "sample", "filter", "aggregate")
 
-`vroom (full altrep)_base` <- function(file, desc) {
-  bench::workout(description = desc,
-    {
-    {library(vroom); Sys.setenv("VROOM_USE_ALTREP_NUMERICS" = "true") }
-      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character())
-      print(x)
-      a <- head(x)
-      b <- tail(x)
-      c <- x[sample(NROW(x), 100), ]
-      d <- x[x$X1 == "helpless_sheep", ]
-      e <- tapply(x$X2, x$X1, function(x) mean(nchar(x)))
-    }
-  )
-}
-
-`vroom (full altrep)_dplyr` <- function(file, desc) {
-  bench::workout(description = desc,
-    {
-      {library(vroom); library(dplyr); Sys.setenv("VROOM_USE_ALTREP_NUMERICS" = "true") }
-      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character())
-      print(x)
-      a <- head(x)
-      b <- tail(x)
-      c <- sample_n(x, 100)
-      d <- filter(x, X1 == "helpless_sheep")
-      e <- group_by(x, X1) %>% summarise(avg_nchar = mean(nchar(X2)))
-    }
-  )
-}
-
 vroom_base <- function(file, desc) {
   bench::workout(description = desc,
     {
@@ -78,11 +50,71 @@ vroom_base <- function(file, desc) {
   )
 }
 
+`vroom (full altrep)_base` <- function(file, desc) {
+  bench::workout(description = desc,
+    {
+    ({library(vroom)})
+      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character(), altrep_opts = TRUE)
+      print(x)
+      a <- head(x)
+      b <- tail(x)
+      c <- x[sample(NROW(x), 100), ]
+      d <- x[x$X1 == "helpless_sheep", ]
+      e <- tapply(x$X2, x$X1, function(x) mean(nchar(x)))
+    }
+  )
+}
+
+`vroom (no altrep)_base` <- function(file, desc) {
+  bench::workout(description = desc,
+    {
+    ({library(vroom)})
+      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character(), altrep_opts = FALSE)
+      print(x)
+      a <- head(x)
+      b <- tail(x)
+      c <- x[sample(NROW(x), 100), ]
+      d <- x[x$X1 == "helpless_sheep", ]
+      e <- tapply(x$X2, x$X1, function(x) mean(nchar(x)))
+    }
+  )
+}
+
 vroom_dplyr <- function(file, desc) {
   bench::workout(description = desc,
     {
-      { library(vroom); library(dplyr) }
+      ({ library(vroom); library(dplyr) })
       x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character())
+      print(x)
+      a <- head(x)
+      b <- tail(x)
+      c <- sample_n(x, 100)
+      d <- filter(x, X1 == "helpless_sheep")
+      e <- group_by(x, X1) %>% summarise(avg_nchar = mean(nchar(X2)))
+    }
+  )
+}
+
+`vroom (full altrep)_dplyr` <- function(file, desc) {
+  bench::workout(description = desc,
+    {
+      ({library(vroom); library(dplyr)})
+      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character(), altrep_opts = TRUE)
+      print(x)
+      a <- head(x)
+      b <- tail(x)
+      c <- sample_n(x, 100)
+      d <- filter(x, X1 == "helpless_sheep")
+      e <- group_by(x, X1) %>% summarise(avg_nchar = mean(nchar(X2)))
+    }
+  )
+}
+
+`vroom (no altrep)_dplyr` <- function(file, desc) {
+  bench::workout(description = desc,
+    {
+      ({library(vroom); library(dplyr)})
+      x <- vroom(file, trim_ws = FALSE, quote = "", escape_double = FALSE, na = character(), altrep_opts = FALSE)
       print(x)
       a <- head(x)
       b <- tail(x)
@@ -111,7 +143,7 @@ data.table <- function(file, desc) {
 readr <- function(file, desc) {
   bench::workout(description = desc,
     {
-    { library(readr); library(dplyr) }
+    ({ library(readr); library(dplyr) })
       x <- read_tsv(file, trim_ws = FALSE, quote = "", na = character())
       print(x)
       a <- head(x)
@@ -126,7 +158,7 @@ readr <- function(file, desc) {
 read.delim <- function(file, desc) {
   bench::workout(description = desc,
     {
-      {}
+      ({})
       x <- read.delim(file, quote = "", na.strings = NULL, stringsAsFactors = FALSE)
       print(head(x, 10))
       a <- head(x)
@@ -139,10 +171,12 @@ read.delim <- function(file, desc) {
 }
 
 times <- list(
-  `vroom (full altrep)_base` = callr::r(`vroom (full altrep)_base`, list(file, desc)),
-  `vroom (full altrep)_dplyr` = callr::r(`vroom (full altrep)_dplyr`, list(file, desc)),
   vroom_base = callr::r(vroom_base, list(file, desc)),
+  `vroom (full altrep)_base` = callr::r(`vroom (full altrep)_base`, list(file, desc)),
+  `vroom (no altrep)_base` = callr::r(`vroom (no altrep)_base`, list(file, desc)),
   vroom_dplyr = callr::r(vroom_dplyr, list(file, desc)),
+  `vroom (full altrep)_dplyr` = callr::r(`vroom (full altrep)_dplyr`, list(file, desc)),
+  `vroom (no altrep)_dplyr` = callr::r(`vroom (no altrep)_dplyr`, list(file, desc)),
   data.table = callr::r(data.table, list(file, desc)),
   readr = callr::r(readr, list(file, desc)),
   read.delim = callr::r(read.delim, list(file, desc))