+++ src/altrep.cc
@@ -1,6 +1,15 @@
 #include <Rcpp.h>
 
 #include "altrep.h"
+#include "vroom_chr.h"
+#include "vroom_date.h"
+#include "vroom_dbl.h"
+#include "vroom_dttm.h"
+#include "vroom_fct.h"
+#include "vroom_int.h"
+#include "vroom_lgl.h"
+#include "vroom_num.h"
+#include "vroom_time.h"
 #include <thread>
 
 using namespace Rcpp;
@@ -12,26 +21,50 @@ void force_materialization(SEXP x) {
 #endif
 }
 
-// [[Rcpp::export]]
-void vroom_materialize(Rcpp::List x) {
+bool vroom_altrep(SEXP x) {
 #ifdef HAS_ALTREP
-  std::vector<std::thread> t;
-  for (int i = 0; i < x.length(); ++i) {
-
-    // First materialize all of the non-character vectors
-    if (TYPEOF(x[i]) == REALSXP || TYPEOF(x[i]) == INTSXP) {
-      t.emplace_back(std::thread([&, i]() { DATAPTR(x[i]); }));
-    }
-  }
-
-  // Then materialize the rest
-  for (int i = 0; i < x.length(); ++i) {
-    if (!(TYPEOF(x[i]) == REALSXP || TYPEOF(x[i]) == INTSXP)) {
-      DATAPTR(x[i]);
-    }
-  }
-  std::for_each(t.begin(), t.end(), std::mem_fn(&std::thread::join));
+  return R_altrep_inherits(x, vroom_chr::class_t) ||
+         R_altrep_inherits(x, vroom_date::class_t) ||
+         R_altrep_inherits(x, vroom_dbl::class_t) ||
+         R_altrep_inherits(x, vroom_dttm::class_t) ||
+         R_altrep_inherits(x, vroom_fct::class_t) ||
+         R_altrep_inherits(x, vroom_int::class_t) ||
+         // R_altrep_inherits(x, vroom_lgl::class_t) ||
+         R_altrep_inherits(x, vroom_num::class_t) ||
+         R_altrep_inherits(x, vroom_time::class_t);
+#else
+  return false;
 #endif
+}
+
+// [[Rcpp::export]]
+SEXP vroom_materialize(SEXP x, bool replace = false) {
+#ifdef HAS_ALTREP
+  for (R_xlen_t i = 0; i < Rf_xlength(x); ++i) {
+
+    SEXP elt = VECTOR_ELT(x, i);
+    // First materialize all of the non-character vectors
+    if (vroom_altrep(elt)) {
+      DATAPTR(elt);
+    }
+  }
+
+  // If replace replace the altrep vectors with their materialized
+  // vectors
+  if (replace) {
+    for (R_xlen_t i = 0; i < Rf_xlength(x); ++i) {
+      SEXP elt = PROTECT(VECTOR_ELT(x, i));
+      if (vroom_altrep(elt)) {
+        SET_VECTOR_ELT(x, i, R_altrep_data2(elt));
+        R_set_altrep_data2(elt, R_NilValue);
+      }
+      UNPROTECT(1);
+    }
+  }
+
+#endif
+
+  return x;
 }
 
 // [[Rcpp::export]]